+++ services/workflowService.js
@@ -908,6 +908,45 @@ function* setNodeInstanceSVNAddress(nodeInstanceId, SVNAddress) {
   return SVNAddress;
 }
 
+function* attributeAddon(workflowId, ibaAttribute) {
+  const _ = yield workFlow_instance_dao.queryById(workflowId, {addon: 1});
+  if (!_) {
+    throwNosuchThisWorkflowNodeInstance();
+  }
+
+  const {addon = {}} = _;
+  Object.assign(addon, ibaAttribute);
+
+  return yield workFlow_instance_dao.update({
+    _id: _._id,
+    upload: {
+      $set: {
+        addon,
+      }
+    }
+  });
+}
+
+function* updateWorkflowItem(workflowId, key, value) {
+  const _ = yield workFlow_instance_dao.queryById(workflowId, {_id: 1});
+  if (!_) {
+    throwNosuchThisWorkflowNodeInstance();
+  }
+
+  if (key !== "name" && key !== "link") {
+    throwNosuchThisWorkflowNodeInstance();
+  }
+
+  return yield workFlow_instance_dao.update({
+    _id: _._id,
+    upload: {
+      $set: {
+        [key]: value,
+      }
+    }
+  });
+}
+
 module.exports = {
   createWorkflowNode,
   createWorkflow,
@@ -918,6 +957,7 @@ module.exports = {
   promoteProcess,
   retroversion,
 
+  attributeAddon,
   appendGoods2Node,
   appendUser2Members,
   cancelOwner,
@@ -938,5 +978,6 @@ module.exports = {
   setTags,
   updateNodeProduceFile,
   updateProcess,
+  updateWorkflowItem,
   workflowMemberList,
 };