+++ tools/webTools/lib/src/initWebPage.js
@@ -65,19 +65,8 @@ window[MODULE].push(`
 
           <div class="dropdown-divider"></div>
 
-          <div class="ibaAttribute">
-            <div class="form-group row">
-              <label class="col-sm-2 col-form-label">链接</label>
-              <div class="col-sm-10">
-                <p class="form-control-static">http://tru.jwis.cn</p>
-                <span class="dataMagager_modify">
-                  <span class="btn btn-outline-warning btn-sm">修改</span>
-                  <span class="btn btn-outline-danger btn-sm">删除</span>
-                </span>
-              </div>
-            </div>
-          </div>
-
+          <div class="ibaAttribute"></div>
+          
           <div class="dropdown-divider"></div>
 
           <div class="collapse" id="dataManager_collapse">